+++ src/reasoning/algorithms/delta_rule_rewrite.rs
@@ -57,14 +57,13 @@ pub fn make_update_sne_programs(program: &Vec<SugaredRule>) -> (Vec<SugaredRule>
         {
             rule.body.iter().enumerate().for_each(|(idx, body_atom)| {
                 let mut new_rule = rule.clone();
-                new_rule.body = new_rule.body;
                 new_rule.body[idx].symbol = format!("{}{}", DELTA_PREFIX, body_atom.symbol);
 
                 let delta_atom = new_rule.body[idx].clone();
                 new_rule.body.remove(idx);
                 new_rule.body.insert(0, delta_atom);
 
-                delta_nonrecursive_program.push(rule.clone());
+                delta_nonrecursive_program.push(new_rule.clone());
             });
         }
     });
@@ -85,25 +84,34 @@ pub fn make_update_sne_programs(program: &Vec<SugaredRule>) -> (Vec<SugaredRule>
                     .body
                     .iter()
                     .enumerate()
-                    .filter(|(_position, new_rule_body_atom)| !idb_relations.contains(&body_atom.symbol))
+                    .filter(|(_position, new_rule_body_atom)| {
+                        let mut new_rule_body_atom_sym = new_rule_body_atom.symbol.as_str();
+
+                        if let Some(undeltaified) = new_rule_body_atom.symbol.strip_prefix(DELTA_PREFIX) {
+                            new_rule_body_atom_sym = undeltaified;
+                        }
+
+                        !idb_relations.contains(new_rule_body_atom_sym)
+                    }
+                        )
                     .collect();
 
                 if edb_body_atoms.len() > 0 {
                     edb_body_atoms
-                        .iter()
+                        .into_iter()
                         .for_each(|(edb_body_atom_idx, new_rule_body_atom)| {
-                            let mut delta_edb_atom = new_rule_body_atom.clone().clone();
-                            delta_edb_atom.symbol = format!("{}{}", DELTA_PREFIX, body_atom.symbol);
+                            let mut delta_edb_atom = new_rule_body_atom.clone();
+                            delta_edb_atom.symbol = format!("{}{}", DELTA_PREFIX, delta_edb_atom.symbol);
 
-                            let mut new_edb_rule = new_rule.clone();
-                            new_edb_rule.body.remove(*edb_body_atom_idx);
-                            new_edb_rule.body.insert(0, delta_edb_atom);
+                            let mut new_edb_rule_one = new_rule.clone();
+                            new_edb_rule_one.body.remove(edb_body_atom_idx);
+                            new_edb_rule_one.body.insert(0, delta_edb_atom.clone());
 
-                            delta_program.push(new_edb_rule);
+                            delta_program.push(new_edb_rule_one);
                         });
-                } else {
-                    delta_program.push(new_rule);
                 }
+
+                delta_program.push(new_rule);
             }
         })
     });
@@ -120,11 +128,9 @@ pub fn deltaify_idb(program: &Vec<SugaredRule>) -> Vec<SugaredRule> {
     return idb_relations
         .into_iter()
         .map(|rule_head| {
-            let delta_string = format!("{}{}", DELTA_PREFIX, rule_head.symbol);
-
             let mut delta_rule = SugaredRule::default();
             let mut delta_rule_head = rule_head.clone();
-            delta_rule_head.symbol = delta_string;
+            delta_rule_head.symbol = format!("{}{}", DELTA_PREFIX, rule_head.symbol);
             delta_rule.head = delta_rule_head;
             delta_rule.body.push(rule_head);